update:抽离可复用mongoose的Schema实例方法到独立的文件
