+++ modules/db/subs.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose')
 const Schema = mongoose.Schema
 const format = require('../../util/format')
+const findAll = require('./common_methods').findAll
 
 const isMail = format.isMail
 const isUserName = format.isUserName
@@ -28,34 +29,7 @@ SubSchema.methods.addSub = async function (sub, cb) {
   })
 }
 
-SubSchema.methods.findAll = function (opt, cb) {
-  let pageIndex = parseInt(opt.pageIndex, 10) || 0
-  let pageSize = parseInt(opt.pageSize, 10) ? parseInt(opt.pageSize, 10) : (parseInt(opt.pageSize, 10) === 0 ? 0 : 10)
-  let collection = opt.col || ''
-
-  if (!collection) {
-    return Promise.reject(new Error('No Collection Specified'))
-  }
-
-  if (pageSize) {
-    return new Promise(function (resolve) {
-      Subs(collection).find({}).skip((pageIndex - 1) * pageSize).limit(pageSize).exec(function (err, result) {
-        if (Array.isArray(result)) {
-          return resolve(result)
-        }
-        return resolve([])
-      })
-    })
-  }
-  return new Promise (function (resolve) {
-    Subs(collection).find({}).exec(function (err, result) {
-      if (Array.isArray(result)) {
-        return resolve(result)
-      }
-      return resolve([])
-    })
-  })
-}
+SubSchema.methods.findAll = findAll
 
 function Subs (name) {
   return mongoose.model(name, SubSchema)
